@@ -1,26 +1,44 @@
 {
-    "name": "react-admin-crm",
-    "version": "3.3.1",
+    "name": "ra-appwrite-demo",
+    "version": "1.0.0",
     "private": true,
+    "type": "module",
     "dependencies": {
-        "@vitejs/plugin-react": "^2.2.0",
-        "date-fns": "^2.19.0",
-        "faker": "~5.4.0",
-        "ra-appwrite": "^1.0.0",
+        "@mui/icons-material": "^7.1.1",
+        "@mui/material": "^7.1.1",
         "appwrite": "^17.0.0",
-        "react": "^18.2.0",
-        "react-admin": "^5.6.0",
-        "react-dom": "^18.2.0",
-        "vite": "^3.2.0"
+        "clsx": "^2.1.1",
+        "data-generator-retail": "^5.0.0",
+        "date-fns": "^3.6.0",
+        "echarts": "^5.6.0",
+        "faker": "~5.4.0",
+        "inflection": "^3.0.0",
+        "ra-appwrite": "^1.0.0",
+        "ra-core": "^5.8.0",
+        "ra-i18n-polyglot": "^5.8.0",
+        "ra-input-rich-text": "^5.8.0",
+        "ra-language-english": "^5.8.0",
+        "ra-language-french": "^5.8.0",
+        "ra-ui-materialui": "^5.8.0",
+        "react": "^19.0.0",
+        "react-admin": "^5.8.0",
+        "react-dom": "^19.0.0",
+        "react-router": "^6.30.1",
+        "react-router-dom": "^6.30.1"
     },
     "devDependencies": {
         "@types/faker": "~5.1.7",
-        "@types/react": "^18.0.0",
-        "@types/react-dom": "^18.0.0",
+        "@types/node": "^20.10.7",
+        "@types/react": "^19.0.0",
+        "@types/react-dom": "^19.0.0",
+        "@vitejs/plugin-react": "^4.3.4",
+        "dotenv": "^16.5.0",
+        "node-appwrite": "^17.0.0",
+        "rollup-preserve-directives": "^1.1.3",
         "source-map-explorer": "^2.0.0",
-        "tsx": "^3.12.2",
+        "tsx": "^4.8.1",
         "typescript": "^5.1.3",
-        "web-vitals": "^1.0.1"
+        "vite": "^6.2.6"
     },
     "scripts": {
         "start": "vite --port 8000",
@@ -28,19 +46,6 @@
         "build": "vite build",
         "test": "vite test",
         "lint": "eslint --fix ./src",
-        "db-seed": "tsx db-seed.ts"
-    },
-    "homepage": ".",
-    "browserslist": {
-        "production": [
-            ">0.2%",
-            "not dead",
-            "not op_mini all"
-        ],
-        "development": [
-            "last 1 chrome version",
-            "last 1 firefox version",
-            "last 1 safari version"
-        ]
+        "db-seed": "tsx setup.ts"
     }
 }
